poetry-demo/pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["amitaiturkel <amitai.turkel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.4.0"
tox = "^4.11.3"
uvicorn = "^0.23.2"
fastapi = "^0.103.1"
nats-python = "^0.8.0"
pip = "^23.2.1"
asyncio-nats-client = "^0.11.5"
nats-py = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
